Calculate USHE Element f_06 (Financial Aid Type)

Calculate USHE Element f_06 (Financial Aid Type)

Usage

f_06(input_df = usheUtils::fake_fin_aid_df, with_intermediates = FALSE)

Arguments

input_df

A Data Frame. Must contain the following data fields: (financial_aid_fund_id).

with_intermediates

Boolean: Option to include intermediate calculated fields.

Value

Original data frame, with USHE data element f_06 appended. Will also return appended intermediate calculated fields, if option is set.

Details

USHE Documentation

ELEMENT NAME: Financial Aid Type

FIELD NAME: f_type_code.

FIELD FORMAT: Varchar, 6 Characters

DEFINITION: The code that defines what type of financial aid the student was awarded. Aid grouping is how aid is classified for both federal and state reporting.

Examples

f_06()
#>    financial_aid_year_id term_id season  ssn student_id financial_aid_fund_id
#> 1                           <NA>   <NA>            <NA>                  <NA>
#> 2                   <NA>           <NA> <NA>       <NA>                  <NA>
#> 3                   <NA>    <NA>   Fall            <NA>                      
#> 4                   <NA>    <NA> Summer <NA>       <NA>                      
#> 5                                Summer <NA>                                 
#> 6                   <NA>         Summer            <NA>                      
#> 7                                  <NA> <NA>                                 
#> 8                   <NA>    <NA> Spring                                      
#> 9                   <NA>    <NA> Spring            <NA>                  <NA>
#> 10                  <NA>           <NA> <NA>                                 
#> 11                  <NA>           <NA> <NA>                             <NA>
#> 12                          <NA>   Fall            <NA>                      
#> 13                                 <NA>            <NA>                  <NA>
#> 14                  <NA>    <NA> Summer <NA>       <NA>                      
#> 15                               Spring <NA>       <NA>                      
#> 16                  <NA>    <NA> Spring                                      
#> 17                  <NA>    <NA>   Fall            <NA>                      
#> 18                  <NA>         Summer                                      
#> 19                               Summer                                  <NA>
#> 20                               Spring                                      
#> 21                                 Fall            <NA>                      
#> 22                               Spring            <NA>                  <NA>
#> 23                                 <NA> <NA>       <NA>                      
#> 24                          <NA>   <NA>            <NA>                  <NA>
#> 25                  <NA>           <NA> <NA>                             <NA>
#> 26                  <NA>         Summer            <NA>                      
#> 27                  <NA>           <NA> <NA>       <NA>                      
#> 28                                 <NA>                                  <NA>
#> 29                          <NA> Summer            <NA>                  <NA>
#> 30                  <NA>    <NA>   <NA>                                      
#>    amount_offered f_type_code f_06
#> 1            <NA>        <NA> <NA>
#> 2            <NA>        <NA> <NA>
#> 3                                 
#> 4            <NA>                 
#> 5            <NA>                 
#> 6                                 
#> 7            <NA>                 
#> 8            <NA>                 
#> 9            <NA>        <NA> <NA>
#> 10                                
#> 11                       <NA> <NA>
#> 12           <NA>                 
#> 13                       <NA> <NA>
#> 14                                
#> 15           <NA>                 
#> 16           <NA>                 
#> 17                                
#> 18           <NA>                 
#> 19           <NA>        <NA> <NA>
#> 20           <NA>                 
#> 21                                
#> 22                       <NA> <NA>
#> 23                                
#> 24                       <NA> <NA>
#> 25           <NA>        <NA> <NA>
#> 26           <NA>                 
#> 27                                
#> 28           <NA>        <NA> <NA>
#> 29           <NA>        <NA> <NA>
#> 30